+++ extensions/remember-this/src/menu.tsx
@@ -11,8 +11,6 @@ if (!fs.existsSync(REMEMBERING_FILE)) {
 
 const preferences = getPreferenceValues<Preferences>();
 const sizeValue = preferences.size;
-const hideNothingTodoTitle = preferences.hideNothingTodoTitle;
-const nothingTodoTitle = hideNothingTodoTitle ? `` : `Woohoo! You got nothing todo!`;
 
 const second = 1000;
 const minute = 60 * second;
@@ -95,7 +93,7 @@ export default function Command() {
     }
     if (sizeValue === "normal") {
       return (
-        <MenuBarExtra icon={Icon.Stars} title={nothingTodoTitle} tooltip="Enjoy the rest of the day off">
+        <MenuBarExtra icon={Icon.Stars} tooltip="Enjoy the rest of the day off">
           <MenuBarExtra.Item icon={Icon.Stars} title={`Woohoo! You got nothing todo!`} />
           <MenuBarExtra.Item
             icon={Icon.CircleProgress100}
@@ -109,7 +107,7 @@ export default function Command() {
     }
     if (sizeValue === "large") {
       return (
-        <MenuBarExtra icon={Icon.Stars} title={nothingTodoTitle} tooltip="Enjoy the rest of the day off">
+        <MenuBarExtra icon={Icon.Stars} title={`Woohoo! You got nothing todo!`} tooltip="Enjoy the rest of the day off">
           <MenuBarExtra.Item icon={Icon.Stars} title={`Woohoo! You got nothing todo!`} />
           <MenuBarExtra.Item
             icon={Icon.CircleProgress100}
@@ -144,7 +142,7 @@ export default function Command() {
     }
     if (sizeValue === "normal") {
       return (
-        <MenuBarExtra icon={Icon.Stars} title={nothingTodoTitle} tooltip="Enjoy the rest of the day off">
+        <MenuBarExtra icon={Icon.Stars} tooltip="Enjoy the rest of the day off">
           <MenuBarExtra.Item icon={Icon.Stars} title={`Woohoo! You got nothing todo!`} />
           <MenuBarExtra.Item
             icon={Icon.CircleProgress100}
@@ -158,7 +156,7 @@ export default function Command() {
     }
     if (sizeValue === "large") {
       return (
-        <MenuBarExtra icon={Icon.Stars} title={nothingTodoTitle} tooltip="Enjoy the rest of the day off">
+        <MenuBarExtra icon={Icon.Stars} title={`Woohoo! You got nothing todo!`} tooltip="Enjoy the rest of the day off">
           <MenuBarExtra.Item icon={Icon.Stars} title={`Woohoo! You got nothing todo!`} />
           <MenuBarExtra.Item
             icon={Icon.CircleProgress100}
@@ -189,7 +187,7 @@ export default function Command() {
     }
     if (sizeValue === "normal") {
       return (
-        <MenuBarExtra icon={Icon.Stars} title={nothingTodoTitle} tooltip="Enjoy the rest of the day off">
+        <MenuBarExtra icon={Icon.Stars} tooltip="Enjoy the rest of the day off">
           <MenuBarExtra.Item icon={Icon.Stars} title={`Woohoo! You got nothing todo!`} />
           <MenuBarExtra.Item
             icon={Icon.CircleProgress100}
@@ -203,7 +201,7 @@ export default function Command() {
     }
     if (sizeValue === "large") {
       return (
-        <MenuBarExtra icon={Icon.Stars} title={nothingTodoTitle} tooltip="Enjoy the rest of the day off">
+        <MenuBarExtra icon={Icon.Stars} title={`Woohoo! You got nothing todo!`} tooltip="Enjoy the rest of the day off">
           <MenuBarExtra.Item icon={Icon.Stars} title={`Woohoo! You got nothing todo!`} />
           <MenuBarExtra.Item
             icon={Icon.CircleProgress100}